services/requirements.txt
# Model Registry, Router, and Auto-Upgrade Services Requirements

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client
httpx==0.25.2

# Redis client
redis==5.0.1

# Prometheus metrics
prometheus-client==0.19.0

# Async support
asyncio-mqtt==0.16.1

# Data validation
email-validator==2.1.0

# JSON handling
orjson==3.9.10

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
